feat(add reporting to community health worker role): add reporting to chw
